Take continuize out of system/datatypes.py
- This is more on the evolution/ side of things, and better aligns with
  my desire to keep System as primarily a data storage, not a behavioral
  class
