+++ pooltool/evolution/__init__.py
@@ -1,3 +1,4 @@
+from pooltool.evolution.continuize import continuize
 from pooltool.evolution.event_based.simulate import simulate as simulate_event_based
 
 simulate = simulate_event_based
